Fix lots of warnings

+++ OgreMain/src/iOS/OgreTimer.cpp
@@ -53,7 +53,8 @@ uint64 Timer::getMilliseconds()
 {
     struct timeval now;
     gettimeofday( &now, NULL );
-    return ( now.tv_sec - start.tv_sec ) * 1000 + ( now.tv_usec - start.tv_usec ) / 1000;
+    return static_cast<uint64>( now.tv_sec - start.tv_sec ) * 1000 +
+           static_cast<uint64>( ( now.tv_usec - start.tv_usec ) / 1000 );
 }
 
 //--------------------------------------------------------------------------------//
@@ -61,7 +62,8 @@ uint64 Timer::getMicroseconds()
 {
     struct timeval now;
     gettimeofday( &now, NULL );
-    return ( now.tv_sec - start.tv_sec ) * 1000000 + ( now.tv_usec - start.tv_usec );
+    return static_cast<uint64>( now.tv_sec - start.tv_sec ) * 1000000 +
+           static_cast<uint64>( now.tv_usec - start.tv_usec );
 }
 
 //-- Common Across All Timers ----------------------------------------------------//
